api/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
landscape-api-py3==0.7.1
requests==2.27.1
termcolor==1.1.0
urllib3==1.26.9
